Move braille offset helper under textUtils
